Updated the .gitignore file. Added the PromiseKit dependency as a submodule. Updated the Travis CI config. Bumped the podspec version. Upgraded the dependency to the latest version. Created a Xcode project just for the framework because Carthage needs it. Created a single framework target and a single framework tests target for all the supported platforms just like PromiseKit is doing. Updated the Example project and it's Podfile.
