+++ .travis.yml
@@ -1,46 +1,37 @@
 language: objective-c
 osx_image: xcode8.3
-podfile: Example/Podfile
 env:
   global:
     - LC_CTYPE=en_US.UTF-8
     - LANG=en_US.UTF-8
-    - WORKSPACE=Example/AwaitKit.xcworkspace
-    - IOS_FRAMEWORK_SCHEME="AwaitKit iOS"
-    - MACOS_FRAMEWORK_SCHEME="AwaitKit macOS"
-    - TVOS_FRAMEWORK_SCHEME="AwaitKit tvOS"
-    - WATCHOS_FRAMEWORK_SCHEME="AwaitKit watchOS"
+    - PROJECT=AwaitKit.xcodeproj
+    - TARGET="AwaitKit"
   matrix:
-    - SCHEME="$WATCHOS_FRAMEWORK_SCHEME" DESTINATION="OS=3.1,name=Apple Watch - 42mm" RUN_TESTS="NO"  POD_LINT="NO" RUN_CODECOV="NO"
-    - SCHEME="$WATCHOS_FRAMEWORK_SCHEME" DESTINATION="OS=2.0,name=Apple Watch - 42mm" RUN_TESTS="NO"  POD_LINT="NO" RUN_CODECOV="NO"
+    - DESTINATION="OS=3.1,name=Apple Watch - 42mm" RUN_TESTS="NO"  RUN_CODECOV="NO"
+    - DESTINATION="OS=2.0,name=Apple Watch - 42mm" RUN_TESTS="NO"  RUN_CODECOV="NO"
 
-    - SCHEME="$TVOS_FRAMEWORK_SCHEME"    DESTINATION="OS=10.0,name=Apple TV 1080p"    RUN_TESTS="YES" POD_LINT="NO" RUN_CODECOV="NO"
+    - DESTINATION="OS=10.0,name=Apple TV 1080p"    RUN_TESTS="YES" RUN_CODECOV="NO"
 
-    - SCHEME="$IOS_FRAMEWORK_SCHEME"     DESTINATION="OS=10.0,name=iPhone 6s"         RUN_TESTS="YES" POD_LINT="NO" RUN_CODECOV="YES"
-    - SCHEME="$IOS_FRAMEWORK_SCHEME"     DESTINATION="OS=9.0,name=iPad 2"             RUN_TESTS="YES" POD_LINT="NO" RUN_CODECOV="YES"
+    - DESTINATION="OS=10.0,name=iPhone 6s"         RUN_TESTS="YES" RUN_CODECOV="YES"
+    - DESTINATION="OS=9.0,name=iPad 2"             RUN_TESTS="YES" RUN_CODECOV="YES"
 
-    - SCHEME="$MACOS_FRAMEWORK_SCHEME"   DESTINATION="arch=x86_64,platform=macOS"     RUN_TESTS="YES" POD_LINT="NO" RUN_CODECOV="YES"
+    - DESTINATION="arch=x86_64,platform=macOS"     RUN_TESTS="YES" RUN_CODECOV="YES"
 script:
   - set -o pipefail
   - xcodebuild -version
 
   # Build Framework in Debug and Run Tests if specified
   - if [ $RUN_TESTS == "YES" ]; then
-      travis_retry xcodebuild -workspace "$WORKSPACE" -scheme "$SCHEME" -destination "$DESTINATION" -configuration Debug GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES ONLY_ACTIVE_ARCH=NO ENABLE_TESTABILITY=YES test | xcpretty;
+      travis_retry xcodebuild -project "$PROJECT" -target "$TARGET" -destination "$DESTINATION" -configuration Debug GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES ONLY_ACTIVE_ARCH=NO ENABLE_TESTABILITY=YES test | xcpretty;
     else
-      travis_retry xcodebuild -workspace "$WORKSPACE" -scheme "$SCHEME" -destination "$DESTINATION" -configuration Debug ONLY_ACTIVE_ARCH=NO build | xcpretty;
+      travis_retry xcodebuild -project "$PROJECT" -target "$TARGET" -destination "$DESTINATION" -configuration Debug ONLY_ACTIVE_ARCH=NO build | xcpretty;
     fi
 
   # Build Framework in Release and Run Tests if specified
   - if [ $RUN_TESTS == "YES" ]; then
-      travis_retry xcodebuild -workspace "$WORKSPACE" -scheme "$SCHEME" -destination "$DESTINATION" -configuration Release GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES ONLY_ACTIVE_ARCH=NO ENABLE_TESTABILITY=YES test | xcpretty;
+      travis_retry xcodebuild -project "$PROJECT" -target "$TARGET" -destination "$DESTINATION" -configuration Release GCC_INSTRUMENT_PROGRAM_FLOW_ARCS=YES GCC_GENERATE_TEST_COVERAGE_FILES=YES ONLY_ACTIVE_ARCH=NO ENABLE_TESTABILITY=YES test | xcpretty;
     else
-      travis_retry xcodebuild -workspace "$WORKSPACE" -scheme "$SCHEME" -destination "$DESTINATION" -configuration Release ONLY_ACTIVE_ARCH=NO build | xcpretty;
-    fi
-
-  # Run `pod lib lint` if specified
-  - if [ $POD_LINT == "YES" ]; then
-      pod lib lint;
+      travis_retry xcodebuild -project "$PROJECT" -target "$TARGET" -destination "$DESTINATION" -configuration Release ONLY_ACTIVE_ARCH=NO build | xcpretty;
     fi
 after_success:
   # Run the `codecov` service if specified
