+++ Cartfile
@@ -1,1 +1,1 @@
-github "mxcl/PromiseKit" ~> 4.2
+github "mxcl/PromiseKit" ~> 4.3
