+++ Cartfile.resolved
@@ -1,0 +1,1 @@
+github "mxcl/PromiseKit" "4.3.2"
