+++ Example/Podfile
@@ -1,38 +1,10 @@
 source 'https://github.com/CocoaPods/Specs.git'
 use_frameworks!
 
-pod 'PromiseKit', '~> 4.2'
-
-target 'AwaitKit iOS' do
-  platform :ios, '8.0'
-end
-
-target 'AwaitKit iOS Tests' do
-  platform :ios, '8.0'
-end
-
-target 'AwaitKit macOS' do
-  platform :osx, '10.10'
-end
-
-target 'AwaitKit macOS Tests' do
-  platform :osx, '10.10'
-end
-
-target 'AwaitKit tvOS' do
-  platform :tvos, '9.0'
-end
-
-target 'AwaitKit tvOS Tests' do
-  platform :tvos, '9.0'
-end
-
-target 'AwaitKit watchOS' do
-  platform :watchos, '2.0'
-end
+pod 'PromiseKit', '~> 4.3'
 
 target 'Example' do
   platform :ios, '8.0'
-  
+
   pod 'AwaitKit', :path => '..'
 end
